fix tests

+++ src/Fiap.Health.Med.Schedule.Manager.Domain/Interfaces/IScheduleRepository.cs
@@ -10,7 +10,7 @@ public interface IScheduleRepository
     Task<(Models.Schedule?, string)> GetScheduleByIdAndDoctorIdAsync(long scheduleId, int doctorId, CancellationToken ct);
     Task<(bool, string)> UpdatescheduleStatusAsync(long scheduleId, EScheduleStatus newStatus, CancellationToken ct);
     Task<int> CreatePendingScheduleAsync(Models.Schedule schedule, CancellationToken cancellationToken);
-    Task<Models.Schedule> GetScheduleByIdAsync(int scheduleId, CancellationToken ct);
+    Task<Models.Schedule> GetScheduleByIdAsync(long scheduleId, CancellationToken ct);
     Task<(bool, string)> DeleteScheduleStatusAsync(long scheduleId, CancellationToken ct);
     Task<int> UpdateScheduleAsync(Models.Schedule schedule, CancellationToken cancellationToken);
 }
